Fixed issue with incorrect price display of first item in a multi-item order.
git-svn-id: http://192.168.2.90/svn/bartsy-venue-android/Venue@171 c3b44e55-a100-42f6-9c86-32dbf579054b

+++ src/com/vendsy/bartsy/venue/model/Item.java
@@ -111,7 +111,7 @@ public class Item {
 		df.setMinimumFractionDigits(0);
 
 		LinearLayout view = (LinearLayout) inflater.inflate(R.layout.item_order, null);
-		((TextView) view.findViewById(R.id.view_order_mini_base_amount)).setText(df.format(Float.parseFloat(getPrice())));
+		((TextView) view.findViewById(R.id.view_order_mini_base_amount)).setText(df.format(Double.parseDouble(getPrice())));
 		((TextView) view.findViewById(R.id.view_order_title)).setText(getTitle());
 		
 		if (has(getOptionsDescription()))
